Added ActionBar menu items. Currently the only thing they do when pressed is display a "Toast" but feel free to remove any (or all) if you don't like the idea. If the little icons don't upload along with this, let me know and I'll post them (You'll know if they didn't upload if Android Studio marks them red.) manually.
Made onClick a bit more flexible to avoid having to manually change things if we decided to change the text on any of the buttons.

Added a Test activity to the ActionBar so you guys would get an idea of how the LocationActivity and YelpActivity that Kevin implemented work. (Note: results are displayed in the command line.)
I'm thinking they should be Services instead of Activities, so maybe I'll convert them over, unless anyone has a better idea?