Test and Actionbar
Added menu options to the Actionbar. "Toasts" are displayed when they
are clicked.
Also made a little Test activity so you guys can see how the Location
and Yelp that Kevin implemented.
Made onClick a bit more flexible so it wouldn't require manual changes
if we changed any of the buttons' text.